requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==4.0.0

sqlalchemy==2.0.23
psycopg2-binary==2.9.9

gunicorn==21.2.0
python-dotenv==1.0.1
